Reset Password and Logout/Re-Login Done

+++ run.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-rm *.class
+if find . -name "*.class" -print | grep -q '.'; then
+    rm *.class
+fi
+
 javac BankSystem.java
 java BankSystem
